Relocate angle classes to a dedicated namespace.
Moved `Degrees` and `Radians` classes into the `OpenCnC.Net.Math.Angles`
namespace for better organization and separation of concerns.

This change improves code structure and makes the purpose of these
classes more explicit.
